updating the models.py

+++ tele/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 # Create your models here.
 class Chat(models.Model):
-    user_id = models.IntegerField(blank=False)
+    user_id = models.BigIntegerField(blank=False)
     top = models.IntegerField(blank=False)
     middle = models.IntegerField(blank=False, default=0)
     bottom = models.IntegerField(blank=False)
